# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Soccer RESTful API Documentation",
        "description": "Dokumentasi API untuk Soccer RESTful API (Laravel) menggunakan L5 Swagger.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/klasemen": {
            "get": {
                "tags": [
                    "Klasemen"
                ],
                "summary": "Menampilkan daftar klasemen",
                "operationId": "077c4e464b93c43a4b6ba0912dbe95cb",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar klasemen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Klasemen"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access Denied"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan pada server."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ada data klasemen ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak ada data klasemen ditemukan"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Klasemen"
                ],
                "summary": "Menambah klasemen baru",
                "operationId": "9c0f76a001ffeacb3d2bcb461153839a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tim_id",
                                    "points",
                                    "wins",
                                    "draws",
                                    "losses",
                                    "goals_for",
                                    "goals_against"
                                ],
                                "properties": {
                                    "tim_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "points": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "wins": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "draws": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "losses": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "goals_for": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "goals_against": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Klasemen berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Klasemen"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access Denied"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan pada server."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Klasemen": {
                "title": "Klasemen",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tim_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "points": {
                        "type": "integer",
                        "example": 30
                    },
                    "wins": {
                        "type": "integer",
                        "example": 10
                    },
                    "draws": {
                        "type": "integer",
                        "example": 5
                    },
                    "losses": {
                        "type": "integer",
                        "example": 3
                    },
                    "goals_for": {
                        "type": "integer",
                        "example": 25
                    },
                    "goals_against": {
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dengan format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}